refactor: seperate data generation to dataset file
